+++ src/Builder/Concerns/LocatesPhpBinary.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Native\Desktop\Builder\Traits;
+namespace Native\Desktop\Builder\Concerns;
 
 trait LocatesPhpBinary
 {
